# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Virtual Lab Studio API
  version: 0.1.0
paths:
...
  /v1/projects/{project_id}/pre-registrations:
    get:
      summary: List Pre Registrations
      operationId: list_pre_registrations_api_v1_projects__project_id__pre_registrations_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PreRegistrationListItem'
                title: Response List Pre Registrations Api V1 Projects  Project Id  Pre Registrations
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      summary: Create Pre Registration
      operationId: create_pre_registration_api_v1_projects__project_id__pre_registrations_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreRegistrationCreateIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreRegistrationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pre-registrations/{prereg_id}:
    get:
      summary: Get Pre Registration
      operationId: get_pre_registration_api_v1_pre_registrations__prereg_id__get
      parameters:
      - name: prereg_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Prereg Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreRegistrationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      summary: Update Pre Registration
      operationId: update_pre_registration_api_v1_pre_registrations__prereg_id__patch
      parameters:
      - name: prereg_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Prereg Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreRegistrationPatchIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreRegistrationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pre-registrations/{prereg_id}/register:
    post:
      summary: Register Pre Registration
      operationId: register_pre_registration_api_v1_pre_registrations__prereg_id__register_post
      parameters:
      - name: prereg_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Prereg Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreRegistrationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/pre-registrations/{prereg_id}/withdraw:
    post:
      summary: Withdraw Pre Registration
      operationId: withdraw_pre_registration_api_v1_pre_registrations__prereg_id__withdraw_post
      parameters:
      - name: prereg_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Prereg Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreRegistrationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/projects/{project_id}/pre-registration-policy:
    patch:
      summary: Set Pre Registration Policy
      operationId: set_pre_registration_policy_api_v1_projects__project_id__pre_registration_policy_patch
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PolicyIn:
      properties:
        pre_registration_required:
          type: boolean
          title: Pre Registration Required
      type: object
      required:
      - pre_registration_required
      title: PolicyIn
    PolicyOut:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        pre_registration_required:
          type: boolean
          title: Pre Registration Required
        warning:
          anyOf:
          - type: string
          - type: 'null'
          title: Warning
      type: object
      required:
      - project_id
      - pre_registration_required
      title: PolicyOut
    PreRegistrationCreateIn:
      properties:
        title:
          type: string
          minLength: 1
          title: Title
        hypothesis:
          type: string
          title: Hypothesis
          default: ''
        protocol:
          type: string
          title: Protocol
          default: ''
        expected_outcomes:
          type: string
          title: Expected Outcomes
          default: ''
        success_criteria:
          type: string
          title: Success Criteria
          default: ''
        analysis_plan:
          type: string
          title: Analysis Plan
          default: ''
        supersedes_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Supersedes Id
        amendment_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Amendment Reason
      type: object
      required:
      - title
      title: PreRegistrationCreateIn
    PreRegistrationListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        version:
          type: integer
          title: Version
        status:
          type: string
          title: Status
        is_active:
          type: boolean
          title: Is Active
        title:
          type: string
          title: Title
        supersedes_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Supersedes Id
        content_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Hash
        registered_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Registered At
        withdrawn_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Withdrawn At
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - project_id
      - version
      - status
      - is_active
      - title
      - supersedes_id
      - content_hash
      - registered_at
      - withdrawn_at
      - created_at
      - updated_at
      title: PreRegistrationListItem
    PreRegistrationOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        version:
          type: integer
          title: Version
        status:
          type: string
          title: Status
        is_active:
          type: boolean
          title: Is Active
        supersedes_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Supersedes Id
        title:
          type: string
          title: Title
        hypothesis:
          type: string
          title: Hypothesis
        protocol:
          type: string
          title: Protocol
        expected_outcomes:
          type: string
          title: Expected Outcomes
        success_criteria:
          type: string
          title: Success Criteria
        analysis_plan:
          type: string
          title: Analysis Plan
        amendment_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Amendment Reason
        content_json:
          additionalProperties: true
          type: object
          title: Content Json
        content_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Content Hash
        registered_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Registered At
        registered_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Registered By
        withdrawn_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Withdrawn At
        withdrawn_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Withdrawn Reason
        created_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        launch_impact:
          anyOf:
          - type: string
          - type: 'null'
          title: Launch Impact
      type: object
      required:
      - id
      - workspace_id
      - project_id
      - version
      - status
      - is_active
      - supersedes_id
      - title
      - hypothesis
      - protocol
      - expected_outcomes
      - success_criteria
      - analysis_plan
      - amendment_reason
      - content_json
      - content_hash
      - registered_at
      - registered_by
      - withdrawn_at
      - withdrawn_reason
      - created_by
      - created_at
      - updated_at
      title: PreRegistrationOut
    PreRegistrationPatchIn:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        hypothesis:
          anyOf:
          - type: string
          - type: 'null'
          title: Hypothesis
        protocol:
          anyOf:
          - type: string
          - type: 'null'
          title: Protocol
        expected_outcomes:
          anyOf:
          - type: string
          - type: 'null'
          title: Expected Outcomes
        success_criteria:
          anyOf:
          - type: string
          - type: 'null'
          title: Success Criteria
        analysis_plan:
          anyOf:
          - type: string
          - type: 'null'
          title: Analysis Plan
        amendment_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Amendment Reason
      type: object
      title: PreRegistrationPatchIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WithdrawIn:
      properties:
        reason:
          type: string
          minLength: 1
          title: Reason
      type: object
      required:
      - reason
      title: WithdrawIn
...